+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# UNRELEASED
+
+  * Feature: Additional information about repository stars.
+
 # 1.2.0 (2014-10-07)
 
   * Feature: Scraping repository language.